% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_model_pipeline.R
\name{CrossValJM_AUC}
\alias{CrossValJM_AUC}
\title{Title}
\usage{
CrossValJM_AUC(original_table, i, df = 3, input_survival_table, splits)
}
\arguments{
\item{splits}{}
}
\value{

}
\description{
Title
}
