% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_model_pipeline.R
\name{CrossValJM_AUC_repeat}
\alias{CrossValJM_AUC_repeat}
\title{Title}
\usage{
CrossValJM_AUC_repeat(summary_table, repeat_time = 5, i = 5)
}
\arguments{
\item{i}{}
}
\value{

}
\description{
Title
}
