% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_model_pipeline.R
\name{survival_risk_AUC}
\alias{survival_risk_AUC}
\title{Title}
\usage{
survival_risk_AUC(summary_table, input_joint_model, survival_table)
}
\arguments{
\item{survival_table}{}
}
\value{

}
\description{
Title
}
